# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fitting Room Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminSetupRequest {
  organisationTradingName: string;
  administratorForenames: string;
  surname: string;
  designation: string;
  username: string;
  password: string;
  retypePassword: string;
  productCode: string;
}

export interface AdminSetupResponse {
  organisationalId: string;
  organisationTradingName: string;
  administratorForenames: string;
  surname: string;
  designation: string;
  username: string;
  message: string;
}

export interface AdminSetupStatus {
  setupComplete: boolean;
}

export type UserRights = (typeof UserRights)[keyof typeof UserRights];

export const UserRights = {
  store_manager: "store_manager",
  store_supervisor: "store_supervisor",
  administrator: "administrator",
} as const;

export interface User {
  id: number;
  userId: string;
  username: string;
  forenames: string;
  surname: string;
  employeeNumber: string;
  rights: UserRights;
  storeBranchCode: string;
  isActive: boolean;
}

export type CreateUserRequestRights =
  (typeof CreateUserRequestRights)[keyof typeof CreateUserRequestRights];

export const CreateUserRequestRights = {
  store_manager: "store_manager",
  store_supervisor: "store_supervisor",
  administrator: "administrator",
} as const;

export interface CreateUserRequest {
  username: string;
  forenames: string;
  surname: string;
  employeeNumber: string;
  rights: CreateUserRequestRights;
  storeBranchCode: string;
  isActive: boolean;
}

export type UpdateUserRequestRights =
  (typeof UpdateUserRequestRights)[keyof typeof UpdateUserRequestRights];

export const UpdateUserRequestRights = {
  store_manager: "store_manager",
  store_supervisor: "store_supervisor",
  administrator: "administrator",
} as const;

export interface UpdateUserRequest {
  username?: string;
  forenames?: string;
  surname?: string;
  employeeNumber?: string;
  rights?: UpdateUserRequestRights;
  storeBranchCode?: string;
  isActive?: boolean;
}

export interface AdminLoginRequest {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  username: string;
  organisationTradingName: string;
  administratorForenames: string;
  surname: string;
  designation: string;
}

export interface ErrorResponse {
  error: string;
}