# lib/api-zod/src/generated/types/createUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fitting Room Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateUserRequestRights } from "./createUserRequestRights";

export interface CreateUserRequest {
  username: string;
  forenames: string;
  surname: string;
  employeeNumber: string;
  rights: CreateUserRequestRights;
  storeBranchCode: string;
  isActive: boolean;
}