# lib/api-zod/src/generated/types/updateUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fitting Room Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateUserRequestRights } from "./updateUserRequestRights";

export interface UpdateUserRequest {
  username?: string;
  forenames?: string;
  surname?: string;
  employeeNumber?: string;
  rights?: UpdateUserRequestRights;
  storeBranchCode?: string;
  isActive?: boolean;
}