# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Fitting Room Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRights } from "./userRights";

export interface User {
  id: number;
  userId: string;
  username: string;
  forenames: string;
  surname: string;
  employeeNumber: string;
  rights: UserRights;
  storeBranchCode: string;
  isActive: boolean;
}